refactor: Extract static IP proxy helper and domain utilities
- Add getStaticIPProxyAgent() to lib/egress/server.ts for consistent
  proxy URL construction across the codebase
- Extract isIpAddress() and isHostUnderDomain() to types/shared/utils/url_utils.ts
  as reusable utilities
- Update mcp_metadata.ts and github.ts internal server to use the new helper
- Simplify WorkspaceResource by using imported utilities instead of local methods
