+++ front/types/shared/utils/url_utils.ts
@@ -95,3 +95,50 @@ export const validateRelativePath = (
     return { valid: false, sanitizedPath: null };
   }
 };
+
+/**
+ * Check if a string is an IP address (IPv4 or IPv6).
+ * Used to reject IP addresses when matching against domain names.
+ */
+export function isIpAddress(host: string): boolean {
+  // IPv4: four groups of 1-3 digits separated by dots, each 0-255
+  const ipv4Match = /^(\d{1,3})\.(\d{1,3})\.(\d{1,3})\.(\d{1,3})$/.exec(host);
+  if (ipv4Match) {
+    const octets = [ipv4Match[1], ipv4Match[2], ipv4Match[3], ipv4Match[4]];
+    const allValid = octets.every((octet) => {
+      const num = parseInt(octet, 10);
+      return num >= 0 && num <= 255;
+    });
+    if (allValid) {
+      return true;
+    }
+  }
+
+  // IPv6: contains colons and only hex digits, colons, dots (for IPv4-mapped)
+  if (host.includes(":")) {
+    if (/^[0-9a-f:.]+$/i.test(host)) {
+      const parts = host.split(":");
+      if (parts.length <= 9) {
+        return true;
+      }
+    }
+  }
+
+  return false;
+}
+
+/**
+ * Check if a host is under a domain.
+ * - Exact match: host === domain
+ * - Subdomain match: host ends with '.' + domain
+ * Both are normalized to lowercase and trailing dots are removed.
+ */
+export function isHostUnderDomain(host: string, domain: string): boolean {
+  const normalizedHost = host.toLowerCase().replace(/\.$/, "");
+  const normalizedDomain = domain.toLowerCase().replace(/\.$/, "");
+
+  return (
+    normalizedHost === normalizedDomain ||
+    normalizedHost.endsWith("." + normalizedDomain)
+  );
+}
